Changed the structure of the project
new project files and new syntax for the main.
