+++ RedNeuronal/lib/nblibraries.properties
@@ -1,4 +1,4 @@
 libs.CopyLibs.classpath=\
     ${base}/CopyLibs/org-netbeans-modules-java-j2seproject-copylibstask.jar
 libs.CopyLibs.displayName=CopyLibs Task
-libs.CopyLibs.prop-version=2.0
+libs.CopyLibs.prop-version=3.0
