+++ RedNeuronal/nbproject/genfiles.properties
@@ -1,8 +1,8 @@
 build.xml.data.CRC32=a1c86d0c
 build.xml.script.CRC32=e8accf88
-build.xml.stylesheet.CRC32=8064a381@1.80.1.48
+build.xml.stylesheet.CRC32=f85dc8f2@1.92.0.48
 # This file is used by a NetBeans-based IDE to track changes in generated files such as build-impl.xml.
 # Do not edit this file. You may delete it but then the IDE will never regenerate such files for you.
 nbproject/build-impl.xml.data.CRC32=a1c86d0c
-nbproject/build-impl.xml.script.CRC32=91b35d35
-nbproject/build-impl.xml.stylesheet.CRC32=830a3534@1.80.1.48
+nbproject/build-impl.xml.script.CRC32=c4946def
+nbproject/build-impl.xml.stylesheet.CRC32=3a2fa800@1.92.0.48
