+++ RedNeuronal/nbproject/private/private.properties
@@ -1,2 +1,2 @@
 compile.on.save=true
-user.properties.file=C:\\Users\\Usuario\\AppData\\Roaming\\NetBeans\\8.2\\build.properties
+user.properties.file=/home/sohrabvafa/snap/netbeans/common/data/11.1/build.properties
