+++ RedNeuronal/nbproject/private/private.xml
@@ -3,8 +3,9 @@
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
         <group>
-            <file>file:/C:/Users/Usuario/Documents/NetBeansProjects/RedNeuronal/src/redneuronal/RedNeuronal.java</file>
-            <file>file:/C:/Users/Usuario/Documents/NetBeansProjects/RedNeuronal/src/redneuronal/Perceptron.java</file>
+            <file>file:/home/sohrabvafa/Downloads/JavaBasicNeuralNetwork-master/RedNeuronal/src/redneuronal/Perceptron.java</file>
+            <file>file:/home/sohrabvafa/Downloads/JavaBasicNeuralNetwork-master/RedNeuronal/src/redneuronal/RedNeuronal.java</file>
+            <file>file:/home/sohrabvafa/Downloads/JavaBasicNeuralNetwork-master/RedNeuronal/src/redneuronal/Neuron.java</file>
         </group>
     </open-files>
 </project-private>
