+++ RedNeuronal/src/redneuronal/Perceptron.java
@@ -1,8 +1,7 @@
 package redneuronal;
 
 import java.util.LinkedList;
-import java.util.List;
-import java.util.Random;
+
 
 public class Perceptron {
     public float[] inputs;
@@ -78,51 +77,3 @@ public class Perceptron {
     }
 }
 
-class Neuron{
-    public float[] inputs, weights;
-    public float bias;
-    public float output;
-    
-    Random rand;
-    
-    public Neuron(){
-        
-    }
-    
-    public Neuron(float[] inputs){
-        this.inputs = inputs;
-        this.weights = new float[inputs.length];
-        
-        rand = new Random();
-        
-        for(int i = 0; i < weights.length; i++){
-            weights[i] = rand.nextFloat()*2-1;
-        }
-        
-        bias = rand.nextFloat()*2-1;
-    }
-    
-    public Neuron(float[] inputs, float[] weights,float bias){
-        this.inputs = inputs;
-        this.weights = weights;
-        this.bias = bias;
-    }
-    
-    public float Calculation(){
-        float result = 0;
-        
-        for(int i = 0; i < inputs.length;i++){
-            result += inputs[i]*weights[i];
-        }
-        
-        float output = activation(result+bias);       
-        this.output = output;
-
-        return output;
-    }
-    
-    float activation(float x){
-        //return(x>=0?1:-1);
-        return(1/(1+(float)(Math.pow(Math.E,-x))));
-    }
-}
